Ya se puede editar usuarios tanto administradores como usuarios normales, pero los usuarios normales aun no pueden eliminar su cuenta, se modificaron las vistas para que los articulos puedan mostrar html pero al agregar un nuevo articulo seria bueno tener un editor de texto para no escribir el html manualmente, tambien me gustaria afinar mas el estilo de las cards que muestran los articulos en pequeño. Tambien dependiendo si es usuario o administrador, te muestra el perfil o el panel de administrador en el navbar

+++ js/user_delete.js
@@ -2,23 +2,29 @@
 
 document.addEventListener('click', e => {
    //saber si el elemento al que se le hizo click y asignarlo a una constante
-   const deleteIcon = e.target.parentElement.className === 'delete-icon' ? true : false;
-  
-   if(deleteIcon){
-      console.log('delete User')      
-      //Obtenr el id del usuraio cortando el contenido del tag id del elemento a
-      let elementId = e.target.parentElement.id;
-      let dash = elementId.indexOf('-');
-      let userId = elementId.slice(dash+1,elementId.length);
-      deleteUser(userId)
-   }  
+   if(e.target.parentElement !== null){
+      const deleteIcon = e.target.parentElement.className === 'delete-icon' ? true : false;
+     
+      if(deleteIcon){
+         e.preventDefault();
+         console.log('delete User')      
+         //Obtenr el id del usuraio cortando el contenido del tag id del elemento a
+         let elementId = e.target.parentElement.id;
+         let dash = elementId.indexOf('-');
+         let userId = elementId.slice(dash+1,elementId.length);
+         const deleteConfirm = confirm('¿Estas seguro que deseas eliminar el usuario?');
+         if(deleteConfirm){
+            deleteUser(userId);
+         }         
+      }  
+   }
 });
 
 let deleteUser = async id =>{
    const peticion = await fetch(`../php/users_controller.php?method=delete&id=${id}`); 
    const resultado = await peticion.json();
    console.log(resultado)
-   if(resultado[1].msg == 'Usuario eliminado'){
+   if(resultado[1]['db-msg'] == 'Usuario eliminado'){
       location.reload();
    }
 }
